# swagger-spec.yml
info:
  title: "Node Express Server"
  version: "1.0.0"
  description: "Node Express Server With User Creation, Authentication, Password Reset using Email Verification and JWT's."
openapi: "3.0.0"
paths:
  "/user/users":
    get:
      description: "Gets all users."
      tags:
        - User
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
              example:
                success: true
                message: "Successfully fetched all users."
                data: []
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/User"
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
...
  "/user/login":
    post:
      description: Logs a user in.
      tags:
        - User
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                usernameOrEmail:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    enum: [success]
                    example: "success"
                  data:
                    type: object
                    properties:
                      description:
                        type: string
                        example: "Successfully logged in."
                      code:
                        type: number
                        enum: [2000]
                      jwtToken:
                        type: string
                        nullable: true
                        example: zqpmxwonceib123
        "422":
          description: Invalid request body
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    enum: [error]
                    example: "error"
                  data:
                    type: object
                    properties:
                      description:
                        type: string
                        example: "Invalid request body."
                      code:
                        type: number
                        enum: [1012]
        "500":
          description: Caught error
          content:
            application/json:
              schema:
                properties:
                  message:
                    type: string
                    enum: [error]
                    example: "error"
                  data:
                    type: object
                    properties:
                      description:
                        type: string
                        example: "An error was caught. Error: something went wrong."
                      code:
                        type: number
                        enum: [1018]
  "/user/reset-password":
    post:
      description: Sends a password reset email.
      tags:
        - User
      requestBody:
        description: "JWT and the new password."
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                password:
                  type: string
            example:
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VybmFtZSI6InVzZXIiLCJpYXQiOjE1OTYwNzQwNzcsImV4cCI6MTU5NjA3NzY3N30.5f3b3b3b3b3b3b3b3b3b3b3b3b"
              password: "newpassw0rd"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  userFound:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: string
                    nullable: true
              example:
                success: true
                userFound: true
                message: "Reset password email sent."
                data: "123objectId321"
        "422":
          description: Invalid request body
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  userFound:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: string
                    nullable: true
              example:
                success: false
                userFound: false
                message: "Invalid request body."
                data: null
        "500":
          description: Caught error
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  userFound:
                    type: boolean
                  message:
                    type: string
                  data:
                    type: string
                    nullable: true
              example:
                success: false
                userFound: false
                message: "An error occurred. Error: something went wrong."
                data: null
...
  "/user/journal/edit":
    post:
      description: "Edits a journals title, entry and/or category."
      tags:
        - Journal
      requestBody:
        description: "The title, entry or category is required."
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                journalId:
                  type: string
                title:
                  type: string
                  nullable: true
                entry:
                  type: string
                  nullable: true
                category:
                  type: string
                  nullable: true
            example:
              userId: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
              journalId: "eyJhbGciOiJIUzI1NiI4234242423423VCJ9"
              title: "New Title"
              entry: "New entry"
              category: "New Category"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    allOf:
                      - $ref: "#/components/schemas/User"
                      - nullable: true
              example:
                success: true
                message: "Journal edited."
                data:
                  {
                    "username": "user1",
                    "email": "email@gmail.com",
                    "journals": [],
                    "journalCategories": [],
                  }
        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "Invalid request body."
                data: null
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
  "/user/journal/journals/{userId}":
    get:
      summary: "Get a users journals."
      tags:
        - Journal
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
          description: "The users document ID"
          example: "666655343535a02523522ac32423"
      responses:
        "200":
          description: "Success"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    allOf:
                      - $ref: "#/components/schemas/User"
              example:
                success: true
                message: "Successfully fetched the journals."
                data:
                  {
                    "username": "user1",
                    "email": "email@gmail.com",
                    "journals": [],
                    "journalCategories": [],
                  }
        "422":
          description: "Invalid request body"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "Invalid request body."
                data: null
        "500":
          description: "Caught error"
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  data:
                    nullable: true
              example:
                success: false
                message: "An error occurred. Error: something went wrong."
                data: null
...
components:
  schemas:
    User:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
          format: email
        journals:
          type: array
          items:
            $ref: "#/components/schemas/Journal"
        journalCategories:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              selected:
                type: boolean
    Journal:
      type: object
      properties:
        title:
          type: string
          example: "Journal 1"
        entry:
          type: string
          example: "Journal 1"
        category:
          type: string
        date:
          type: string
        selected:
          type: boolean
...